pyproject.toml
[tool.poetry]
name = "unet-sample"
version = "0.1.0"
description = ""
authors = ["hichimada <60519909+Nao-Y1996@users.noreply.github.comi>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
torch = "^2.3.0"
torchvision = "^0.18.0"
opencv-python = "^4.9.0.80"
pillow = "^10.3.0"
numpy = "^1.26.4"
matplotlib = "^3.8.4"
tensorboard = "^2.16.2"
fiftyone = "^0.23.8"
tqdm = "^4.66.2"
jupyter = "^1.0.0"
pycocotools = "^2.0.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
